clean up app
