+++ .github/pull_request_template.md
@@ -10,8 +10,6 @@
 
 -
 
-**Tests and Deployment**
-- [ ] Deploy Stage
 
 **Steps to consider while deploying**
 - Configuration changes:
