+++ .github/workflows/deploy_prod.yml
@@ -23,4 +23,5 @@ jobs:
             -d '{
             "ref": "${{ github.event.inputs.tag_name }}",
             "environment": "production",
-            "auto_merge": false}'
+            "auto_merge": false,
+            "description": "${{secrets.DEPLOYMENT_INFO}}"}'
