+++ .github/workflows/deploy_stage.yml
@@ -23,4 +23,5 @@ jobs:
             -d '{
             "ref": "${{ github.event.inputs.branch_name }}",
             "environment": "stage",
-            "auto_merge": false}'
+            "auto_merge": false,
+            "description": "${{secrets.DEPLOYMENT_INFO}}"}'
