+++ cgdeployment/config.py
@@ -5,9 +5,7 @@ from pydantic import BaseSettings
 
 class EnvConfig(BaseSettings):
     authorization_token: Optional[str]
-    deployments_post_uri: Optional[str]
-    deployment_statuses_post_uri: Optional[str]
-    deployment_diff_trigger: Optional[str]
+    webhook_token: Optional[str]
 
     class Config:
         env_file = ".env"
